+++ src/main/java/blackjack/domain/player/GamePlayer.java
@@ -2,6 +2,9 @@ package blackjack.domain.player;
 
 
 import blackjack.domain.card.Cards;
+import blackjack.domain.player.info.BettingMoney;
+import blackjack.domain.player.info.PlayerInfo;
+import blackjack.domain.player.info.Name;
 
 import java.util.Objects;
