+++ src/main/java/blackjack/domain/player/info/Name.java
@@ -1,4 +1,4 @@
-package blackjack.domain.player;
+package blackjack.domain.player.info;
 
 import java.util.Objects;
