+++ src/main/java/blackjack/domain/player/info/Names.java
@@ -1,4 +1,4 @@
-package blackjack.domain.player;
+package blackjack.domain.player.info;
 
 import java.util.HashSet;
 import java.util.List;
