+++ src/main/java/blackjack/domain/result/DealerResult.java
@@ -1,6 +1,7 @@
 package blackjack.domain.result;
 
-import blackjack.domain.player.Name;
+import blackjack.domain.player.info.Name;
+import blackjack.domain.result.prize.PrizeMoney;
 
 import java.util.List;
