+++ src/main/java/blackjack/view/input/BlackjackCommand.java
@@ -1,4 +1,4 @@
-package blackjack.view;
+package blackjack.view.input;
 
 import java.util.Arrays;
