+++ src/test/java/blackjack/domain/result/DealerResultTest.java
@@ -3,18 +3,14 @@ package blackjack.domain.result;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 
-import blackjack.domain.player.Name;
+import blackjack.domain.player.info.Name;
 
-import java.util.Arrays;
 import java.util.List;
-import java.util.stream.Stream;
 
+import blackjack.domain.result.prize.PrizeMoney;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.MethodSource;
 
 class DealerResultTest {
     private Name dealerName;
