+++ src/test/java/blackjack/fixture/CardFixture.java
@@ -4,7 +4,7 @@ import blackjack.domain.card.Card;
 import blackjack.domain.card.CardSymbol;
 import blackjack.domain.card.CardValue;
 import blackjack.domain.card.Cards;
-import blackjack.domain.card.Deck;
+import blackjack.domain.blackjack.Deck;
 
 import java.util.ArrayDeque;
 import java.util.List;
