+++ src/test/java/blackjack/fixture/ParticipantImpl.java
@@ -1,9 +1,7 @@
 package blackjack.fixture;
 
 import blackjack.domain.card.Cards;
-import blackjack.domain.player.Name;
 import blackjack.domain.player.Participant;
-import blackjack.domain.player.PlayerInfo;
 
 public class ParticipantImpl extends Participant {
     @Override
