+++ src/test/java/blackjack/fixture/PlayerFixture.java
@@ -3,7 +3,7 @@ package blackjack.fixture;
 import blackjack.domain.card.CardValue;
 import blackjack.domain.player.Dealer;
 import blackjack.domain.player.GamePlayer;
-import blackjack.domain.player.Name;
+import blackjack.domain.player.info.Name;
 import blackjack.domain.player.Participant;
 
 import java.util.List;
